refactor: keep runtime-use factory internal
